+++ packages/kiarina-lib-cloudflare-d1/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Refactored internal module structure for better organization
+- **BREAKING**: Renamed function parameters for consistency
+  - `config_key` → `settings_key` in `create_d1_client()`
+  - `auth_config_key` → `auth_settings_key` in `create_d1_client()`
+
 ## [1.11.2] - 2025-12-02
 
 ### Changed
