+++ packages/kiarina-lib-cloudflare-d1/src/kiarina/lib/cloudflare/d1/asyncio.py
@@ -3,14 +3,15 @@ from importlib import import_module
 from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
-    from ._async.client import D1Client
-    from ._async.registry import create_d1_client
+    from ._async.helpers.create_d1_client import create_d1_client
+    from ._async.models.d1_client import D1Client
     from .settings import D1Settings, settings_manager
 
 __all__ = [
-    # ._async
-    "D1Client",
+    # ._async.helpers
     "create_d1_client",
+    # ._async.models
+    "D1Client",
     # .settings
     "D1Settings",
     "settings_manager",
@@ -24,9 +25,10 @@ def __getattr__(name: str) -> object:
         raise AttributeError(f"module '{__name__}' has no attribute '{name}'")
 
     module_map = {
-        # ._async
-        "D1Client": "._async.client",
-        "create_d1_client": "._async.registry",
+        # ._async.helpers
+        "create_d1_client": "._async.helpers.create_d1_client",
+        # ._async.models
+        "D1Client": "._async.models.d1_client",
         # .settings
         "D1Settings": ".settings",
         "settings_manager": ".settings",
